Change to common dependencies

+++ build_scripts/dd-module.sh
@@ -1,8 +1,22 @@
 #!/bin/bash
 
+##Set branch to pull from for dependencies
+set -ea
+
+: "${DEPENDENCY_BRANCH:=master}"
+
 ##Pull variables from github
-wget -nc https://raw.githubusercontent.com/CHBMB/Unraid-DVB/master/build_scripts/variables.sh
-. "$(dirname "$(readlink -f ${BASH_SOURCE[0]})")"/variables.sh
+wget -nc https://raw.githubusercontent.com/linuxserver/Unraid-Dependencies/${DEPENDENCY_BRANCH}/build_scripts/variables.sh
+wget -nc https://raw.githubusercontent.com/linuxserver/Unraid-Dependencies/${DEPENDENCY_BRANCH}/build_scripts/dvb-variables.sh
+
+source ./variables.sh
+
+if [[ -z "$D" ]]; then
+    echo "Must provide D in environment" 1>&2
+    exit 1
+fi
+
+source ${D}/dvb-variables.sh
 
 ##Restore /lib/modules/
 rm -rf  /lib/modules
@@ -24,7 +38,7 @@ make install
 
 #Create /etc/unraid-media to identify type of mediabuild and copy to bzroot
 echo base=\"Digital Devices \(Github\)\" > /lib/firmware/unraid-media
-echo driver=\"${DATE}\" >> /lib/firmware/unraid-media
+echo driver=\"${DD}\" >> /lib/firmware/unraid-media
 
 #Copy /etc/unraid-media to identify type of mediabuild to destination folder
 mkdir -p ${D}/${UNRAID_VERSION}/dd/
